// src/types/data-grid.ts
export type RowHeightValue = "short" | "medium" | "tall" | "extra-tall"
// ...
export interface CellUpdate {
  rowIndex: number
  columnId: string
  value: unknown
}
// ...
  interface TableMeta<TData extends RowData> {
    customCell?: (props: DataGridCellProps<TData>) => React.ReactNode
    dataGridRef?: React.RefObject<HTMLElement | null>
    cellMapRef?: React.RefObject<Map<string, HTMLDivElement>>
    focusedCell?: CellPosition | null
    editingCell?: CellPosition | null
    selectionState?: SelectionState
    searchOpen?: boolean
    getIsCellSelected?: (rowIndex: number, columnId: string) => boolean
    getIsSearchMatch?: (rowIndex: number, columnId: string) => boolean
    getIsActiveSearchMatch?: (rowIndex: number, columnId: string) => boolean
    getVisualRowIndex?: (rowId: string) => number | undefined
    rowHeight?: RowHeightValue
    onRowHeightChange?: (value: RowHeightValue) => void
    onRowSelect?: (rowId: string, checked: boolean, shiftKey: boolean) => void
    onDataUpdate?: (params: CellUpdate | Array<CellUpdate>) => void
    onRowsDelete?: (rowIndices: number[]) => void | Promise<void>
    onColumnClick?: (columnId: string) => void
    onCellClick?: (
      rowIndex: number,
      columnId: string,
      event?: React.MouseEvent
    ) => void
    onCellDoubleClick?: (rowIndex: number, columnId: string) => void
    onCellMouseDown?: (
      rowIndex: number,
      columnId: string,
      event: React.MouseEvent
    ) => void
    onCellMouseEnter?: (rowIndex: number, columnId: string) => void
    onCellMouseUp?: () => void
    onCellContextMenu?: (
      rowIndex: number,
      columnId: string,
      event: React.MouseEvent
    ) => void
    onCellEditingStart?: (rowIndex: number, columnId: string) => void
    onCellEditingStop?: (opts?: {
      direction?: NavigationDirection
      moveToNextRow?: boolean
    }) => void
    onCellsCopy?: () => void
    onCellsCut?: () => void
    onCellsPaste?: (expand?: boolean) => void
    onSelectionClear?: () => void
    onFilesUpload?: (params: {
      files: File[]
      rowIndex: number
      columnId: string
    }) => Promise<FileCellData[]>
    onFilesDelete?: (params: {
      fileIds: string[]
      rowIndex: number
      columnId: string
    }) => void | Promise<void>
    contextMenu?: ContextMenuState
    onContextMenuOpenChange?: (open: boolean) => void
    pasteDialog?: PasteDialogState
    onPasteDialogOpenChange?: (open: boolean) => void
    readOnly?: boolean
  }
// ...
export interface CellPosition {
  rowIndex: number
  columnId: string
}

export interface CellRange {
  start: CellPosition
  end: CellPosition
}

export interface SelectionState {
  selectedCells: Set<string>
  selectionRange: CellRange | null
  isSelecting: boolean
}

export interface ContextMenuState {
  open: boolean
  x: number
  y: number
}

export interface PasteDialogState {
  open: boolean
  rowsNeeded: number
  clipboardText: string
}

export type NavigationDirection =
  | "up"
  | "down"
  | "left"
  | "right"
  | "home"
  | "end"
  | "ctrl+up"
  | "ctrl+down"
  | "ctrl+home"
  | "ctrl+end"
  | "pageup"
  | "pagedown"
  | "pageleft"
  | "pageright"

export interface SearchState {
  searchMatches: CellPosition[]
  matchIndex: number
  searchOpen: boolean
  onSearchOpenChange: (open: boolean) => void
  searchQuery: string
  onSearchQueryChange: (query: string) => void
  onSearch: (query: string) => void
  onNavigateToNextMatch: () => void
  onNavigateToPrevMatch: () => void
}

export interface DataGridCellProps<TData> {
  cell: Cell<TData, unknown>
  tableMeta: TableMeta<TData>
  rowIndex: number
  columnId: string
  rowHeight: RowHeightValue
  isEditing: boolean
  isFocused: boolean
  isSelected: boolean
  isSearchMatch: boolean
  isActiveSearchMatch: boolean
  readOnly: boolean
  renderValue?: () => unknown
}

declare module "@tanstack/react-table" {
  interface ColumnMeta<TData, TValue> {
     value?: TValue
    customCell?: (props: DataGridCellProps<TData>) => React.ReactNode
  }
}
export interface FileCellData {
  id: string
  name: string
  size: number
  type: string
  url?: string
}